fix(vitest): move internal vitest types to .d.ts
